requirements.txt

fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
psycopg2-binary
